// local-hobbies-frontend/src/types/user.ts
export enum UserGenderMatched {
    Man = 'Man',
    Nonbinary = 'Nonbinary',
    Woman = 'Woman',
}

/**
 * Name of onboarding section, used to indicate which sections are incomplete
 */
export enum OnboardingSectionName {
    Name = 'name',
    BirthDate = 'birthDate',
    Location = 'location',
    PublicContactInfo = 'publicContactInfo',
    GenderMatched = 'genderMatched',
    ShowAge = 'showAge',
    ShowGenderDisplayed = 'showGenderDisplayed',
    Hobbies = 'hobbies',
    Availabilities = 'availabilities',
}
// ...
export interface UserOnboardingRequest {
    name?: string;
    /** ISO 8601 date (YYYY-MM-DD) */
    birthDate?: string;
    location?: GeoJsonPoint;
    publicContactInfo?: string;
    genderMatched?: UserGenderMatched;
    showAge?: boolean;
    showGenderDisplayed?: boolean;
    hobbies?: HobbyCreationRequest[];
    availabilities?: AvailabilityOnboardingRequests;
}